split up the constants file
