+++ src/client/app/reducers/editors.jsx
@@ -1,43 +1,11 @@
 import { EditorState, convertFromRaw } from 'draft-js';
 import * as ActionTypes from '../constants.jsx';
+import * as Code from '../codeConstants.jsx';
 
 const initialState = {
   editors: {},
   editorIndex: 0
 };
-
-
-const defaultHTML =
-`<!DOCTYPE html>
-<html>
-  <head>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/p5.js/0.5.11/p5.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/p5.js/0.5.11/addons/p5.dom.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/p5.js/0.5.11/addons/p5.sound.min.js"></script>
-    <meta charset="utf-8" />
-  </head>
-  <body>
-    <script>
-      function setup() {
-        createCanvas(400, 400);
-      }
-
-      function draw() {
-        background(220);
-      }
-    </script>
-  </body>
-</html>
-
-`;
-
-const defaultSketch = `function setup() {
-  createCanvas(400, 400);
-  console.log("drawing a canvas");
-}
-function draw() {
-  background(220);
-}`;
 
 let stack = [];
 
@@ -110,12 +78,7 @@ const editorsReducer = (state = initialState, action) => {
         id,
         index: stack.length,
         consoleOutputText: [],
-        files: [
-          {
-            name: 'index.html',
-            content: defaultHTML
-          }
-        ],
+        files: Code.FILES.p5,
         isPlaying: false,
         editorMode: 'p5',
         x: 0,
@@ -158,7 +121,6 @@ const editorsReducer = (state = initialState, action) => {
         editors
       });
     }
-
 
     /** TEXT EDITOR */
     case ActionTypes.ADD_TEXT_EDITOR: {
